Refactored function with multiple declarations
